requirements.txt
Pillow==2.6.1
oauthlib==0.7.2
redis==2.10.3
requests==2.4.3
requests-oauthlib==0.4.1
six==1.7.3
tweepy==3.1.0
wsgiref==0.1.2
